backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn
sqlalchemy
alembic
psycopg[binary]
pydantic
pydantic-settings
python-jose[cryptography]
bcrypt
python-multipart
email-validator
orjson
httpx
pytest
pytest-asyncio
pytest-cov
